@@ -1,44 +1,34 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Reflection;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
+﻿using System.Reflection;
 
-namespace ThMouseX
+namespace ThMouseX;
+
+public partial class AboutForm : Form
 {
-    public partial class AboutForm : Form
+    public static readonly Icon AppIcon = Icon.ExtractAssociatedIcon(Assembly.GetExecutingAssembly().Location);
+    public static readonly Image Logo = AppIcon.ToBitmap();
+    public AboutForm()
     {
-        public static readonly Icon AppIcon = Icon.ExtractAssociatedIcon(Assembly.GetExecutingAssembly().Location);
-        public static readonly Image Logo = AppIcon.ToBitmap();
-        public AboutForm()
-        {
-            InitializeComponent();
-            Icon = AppIcon;
-            imgMouse.Image = Logo;
-            Text = string.Format(Text, Program.AppName);
-            lblDescription1.Text = string.Format(lblDescription1.Text, Program.AppName);
-            lblDescription2.Text = string.Format(lblDescription2.Text, Program.AppName);
-        }
+        InitializeComponent();
+        Icon = AppIcon;
+        imgMouse.Image = Logo;
+        Text = string.Format(Text, Program.AppName);
+        lblDescription1.Text = string.Format(lblDescription1.Text, Program.AppName);
+        lblDescription2.Text = string.Format(lblDescription2.Text, Program.AppName);
+    }
 
-        private void AboutForm_Shown(object sender, EventArgs e)
-        {
-            ShowIcon = false;
-        }
+    private void AboutForm_Shown(object sender, EventArgs e)
+    {
+        ShowIcon = false;
+    }
 
-        private void cmdOK_Click(object sender, EventArgs e)
-        {
-            Close();
-        }
+    private void cmdOK_Click(object sender, EventArgs e)
+    {
+        Close();
+    }
 
-        private void cmdQuit_Click(object sender, EventArgs e)
-        {
-            DialogResult = DialogResult.Abort;
-            Close();
-        }
+    private void cmdQuit_Click(object sender, EventArgs e)
+    {
+        DialogResult = DialogResult.Abort;
+        Close();
     }
 }
