@@ -1,49 +1,45 @@
-﻿using System;
-using System.Runtime.InteropServices;
-using System.Threading;
-using System.Windows.Forms;
+﻿using System.Runtime.InteropServices;
 
-namespace ThMouseX
+namespace ThMouseX;
+
+static class Program
 {
-    static class Program
+    public const string AppName = "ThMouseX";
+    public const string DllName = AppName + ".dll";
+
+    [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
+    static extern bool InstallHooks();
+    [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
+    static extern void RemoveHooks();
+    [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
+    static extern bool ReadGamesFile();
+    [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
+    static extern bool ReadGeneralConfigFile();
+
+    [STAThread]
+    static void Main()
     {
-        public const string AppName = "ThMouseX";
-        public const string DllName = AppName + ".dll";
+        Application.EnableVisualStyles();
+        Application.SetCompatibleTextRenderingDefault(false);
 
-        [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
-        static extern bool InstallHooks();
-        [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
-        static extern void RemoveHooks();
-        [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
-        static extern bool ReadGamesFile();
-        [DllImport(DllName, CallingConvention = CallingConvention.Cdecl)]
-        static extern bool ReadGeneralConfigFile();
-
-        [STAThread]
-        static void Main()
+        using var mutex = new Mutex(true, AppName, out var mutexIsCreated);
+        if (!mutexIsCreated)
         {
-            Application.EnableVisualStyles();
-            Application.SetCompatibleTextRenderingDefault(false);
-
-            using var mutex = new Mutex(true, AppName, out var mutexIsCreated);
-            if (!mutexIsCreated)
-            {
-                MessageBox.Show($"{AppName} is already running.", AppName, MessageBoxButtons.OK, MessageBoxIcon.Information);
-                Environment.Exit(1);
-            }
-
-            if (!ReadGamesFile())
-                Environment.Exit(1);
-
-            if (!ReadGeneralConfigFile())
-                Environment.Exit(1);
-
-            if (!InstallHooks())
-                Environment.Exit(1);
-
-            Application.Run(new ThMouseApplicationContext());
-
-            RemoveHooks();
+            MessageBox.Show($"{AppName} is already running.", AppName, MessageBoxButtons.OK, MessageBoxIcon.Information);
+            Environment.Exit(1);
         }
+
+        if (!ReadGamesFile())
+            Environment.Exit(1);
+
+        if (!ReadGeneralConfigFile())
+            Environment.Exit(1);
+
+        if (!InstallHooks())
+            Environment.Exit(1);
+
+        Application.Run(new ThMouseApplicationContext());
+
+        RemoveHooks();
     }
 }
